Add logic & Fix bugs
- 리마인더를 리스트뷰 형태로 확인 할 수 있도록 추가
- Realm 데이터 스키마 문제 해결

+++ app/src/main/java/com/tedkim/smartschedule/calendar/ScheduleListRealmAdapter.java
@@ -113,7 +113,7 @@ public class ScheduleListRealmAdapter extends RealmRecyclerViewAdapter<ScheduleD
         });
     }
 
-    private void setFragmentDialog(int position){
+    private void setFragmentDialog(long position){
 
         FragmentManager fragmentManager = ((AppCompatActivity)mContext).getSupportFragmentManager();
         FragmentTransaction transaction = fragmentManager.beginTransaction();
